# backend/cmd/server/docs/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "description": "登录后返回的 JWT，在 Authorize 中填入 token 即可",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "告警路由与通知平台 API。需先调用 POST /api/v1/auth/login 获取 token，再在请求头中携带 Authorization: Bearer <token> 调用其他接口。",
    "title": "KK Alert API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/datasources/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "删除数据源"
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "datasource"
          }
        },
        "summary": "数据源详情"
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "更新数据源"
      }
    },
...
    "/api/v1/rules/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "删除规则"
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "rule"
          }
        },
        "summary": "规则详情"
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "更新规则"
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Current host",
      "url": "/"
    }
  ]
}